Properly implement payment_queryInfo (#1098)
* Properly implement payment_queryInfo

* Finish change

* Dummy commit

+++ src/json_rpc/methods.rs
@@ -500,7 +500,7 @@ pub struct RuntimeVersion {
 pub struct RuntimeDispatchInfo {
     pub weight: u64,
     pub class: DispatchClass,
-    pub partial_fee: u64,
+    pub partial_fee: u128,
 }
 
 #[derive(Debug, Copy, Clone)]
